requirements.txt
# Core Framework
selenium>=4.28.0
pytest>=8.3.0

# Reporting & Visibility
pytest-html~=4.1.0
allure-pytest>=2.13.5
pytest-metadata>=3.1.0

# Execution & Stability
pytest-xdist>=3.6.0
pytest-rerunfailures>=15.0

# Utilities & Data
python-dotenv>=1.0.1
openpyxl>=3.1.5
requests>=2.32.0